leds: LP5562: Fix wait command misinterpreted as ramp
After branching into a loop, the 'lsb' variable was filled with the jump
programm counter, and I forgot to clear the variable when queuing the
wait command, effectively turning the wait command into a ramp,
causing the green channel (Because we're using that as the wait master)
to slowly light up with each new cycle if the delay cycle and jump address
machted the bits of the ramp command.

Since a wait command has an LSB of 0, just hardcode it.

Change-Id: Iab325c7f493a13e9a34517d5bc50ef5b2930f2bc
Signed-off-by: Andrei F <luxneb@gmail.com>